+++ tests/testthat/test-power4test_print_indirect1.R
@@ -83,7 +83,9 @@ power_all_sim_only <- power4test(nrep = 10,
 
 expect_output(print(power_all_sim_only),
               "x -> m1 -> m2")
-tmp <- pop_indirect(power_all_sim_only$sim_all)
+tmp <- pop_indirect(power_all_sim_only$sim_all,
+                    pure_x = FALSE,
+                    pure_y = FALSE)
 expect_equal(coef(tmp)[3],
              .141,
              ignore_attr = TRUE)